+++ test/amp/StableSortTest/StableSortTest.cpp
@@ -1919,11 +1919,11 @@ public:
 INSTANTIATE_TEST_CASE_P(sortDescending, withStdVect, ::testing::Range(50, 100, 1));
 
 TEST (sanity_sort__withBoltClDevVectDouble_epr, floatSerial){
-    size_t sizeOfInputBufer = 64; //test case is failing for all values greater than 32
+    int sizeOfInputBufer = 64; //test case is failing for all values greater than 32
     std::vector<double>  stdVect(0);
     bolt::amp::device_vector<double>  boltVect(0);
 
-    for (size_t i = 0 ; i < sizeOfInputBufer; i++){
+    for (int i = 0 ; i < sizeOfInputBufer; i++){
         double dValue = rand();
         dValue = dValue/rand();
         dValue = dValue*rand();
